+++ src/keywords.ts
@@ -1,0 +1,29 @@
+import * as vscode from 'vscode';
+
+export const keywords = [
+    new vscode.CompletionItem("function", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("and", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("break", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("do", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("else", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("elseif", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("end", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("for", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("goto", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("if", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("in", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("local", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("not", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("or", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("repeat", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("return", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("then", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("until", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("while", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("switch", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("case", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("continue", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("defer", vscode.CompletionItemKind.Keyword),
+    new vscode.CompletionItem("global", vscode.CompletionItemKind.Keyword),
+]
+